Worked on shit

+++ Concrete Table/src/dto/MetalDTO.java
@@ -3,11 +3,11 @@ package dto;
 public class MetalDTO {
     public long id;
     public String name;
-    public long atomicNumber;
-    public long atomicMass;
+    public int atomicNumber;
+    public double atomicMass;
     public long dissolvedBy;
 
-    public MetalDTO(long id, String name, long atomicNumber, long atomicMass, long dissolvedBy) {
+    public MetalDTO(long id, String name, int atomicNumber, double atomicMass, long dissolvedBy) {
         this.id = id;
         this.name = name;
         this.atomicNumber = atomicNumber;
